Fix multiple issued with pathing
This patch firstly fixes a few errors is in the *re-implementation* of the
algorithms. There were a few errors which caused mainly the min-heap and A*
algorithms to produce slightly incorrect paths.

These errors were located in the implementation of MazeVertex. Where there were
two main issues. Firstly, the "makeNext(x, y, end)" method did not properly add
"+1" to the new vertex. Which meant that each new Vertext in the "chain" had the
same length as its parent, which of course messed up the pathing.

Secondly, the "proximityTo" method was incorrect, and produced bad results due
to a missunderstanding of how Java treads the order-of-operations with math.

Then lastly, the MazeReader was modified to premtively disable the "block size"
and instead use all the pixels within the "maze area" as cells; in a so called
cell-per-pixel configuration. This is significantly slower, since it
exponentially increases the number of cells. But the advantage is that with the
additional cells the differences between the algorithms are highlighted.

With the previous method (which is so much better, in terms of speed) however,
since the reader "scaled down" the maze, it looked like all the algorithms found
the exact same path. Which is technically true, but the only reason for this is
because the effective size of the maze was reduced so much, that all the
algorithms "agreed" on the path. But as mentioned, this meant that you couldn't
tell them apart. So this feature was removed.

This patch is only part of the process. A future commit will strip out the
remaining "dead code".