Strip canned "blockSize" and "borderThickness"
This commit removes the blockSize and borderThickness features from the Model,
as it was deemed disadvantageous.

These parts made the algorithms much more efficient, but at the cost of making
it so that the differences between the algorithms was obscured.

Is any real world application, these functions would have stayed since the point
would be maximum performance. But as this is an assignment, where at least in
part the display of difference is the goal, it had to go.