Fix remove unused calculation in Dijkstra's Queue
In the Dijkstra's Queue implementation when creating a new vertex using
the *MazeVertex.makeNext()* method there is an option to provide a point
which is treated as the "end goal" of the maze.

If this point is provided, the new MazeVertex's "directDistance"
property will automatically be calculated and stored during
initialization.

However, in the Dijkstra's Queue implementation, this is not necessary
since the implementation at no point makes use of the "directDistance"
value. Hence this fix removes this.